Create collateralized loan default simulation with jump-diffusion process
Assume that the value of a collateral follows a jump-diffusion process:
𝑑𝑉𝑡𝑉𝑡−=𝜇𝑑𝑡+𝜎𝑑𝑊𝑡+𝛾𝑑𝐽𝑡
where 𝜇,𝜎,𝛾<0, and 𝑉0 are given, 𝐽 is a Poisson process, with intensity 𝜆1, independent of the Brownian Motion process 𝑊.
𝑉𝑡− is the value process before jump occurs at time t (if any).
Consider a collateralized loan, with a contract rate per period r and maturity T on the above-collateral, and assume the outstanding balance of that loan follows this process:
𝐿𝑡=𝑎−𝑏𝑐12𝑡
where 𝑎>0,𝑏>0 ,𝑐>1,𝑎𝑛𝑑 𝐿0 are given. We have that 𝐿𝑇=0.
Define the following stopping time: 𝑄=𝑚𝑖𝑛{𝑡≥0: 𝑉𝑡≤𝑞𝑡𝐿𝑡}
This stopping time is the first time when the relative value of the collateral (with respect to the outstanding loan balance) crosses a threshold which will be viewed as the “optimal exercise boundary” of the option to default.
Define another stopping time, which is the first time an adverse event occurs:
𝑆=𝑚𝑖𝑛{𝑡≥0: 𝑁𝑡>0}
Assume that 𝑁𝑡 is a Poisson process with intensity of 𝜆2.
Define 𝜏=min{𝑄,𝑆}.
We assume the embedded default option will be exercised at time 𝜏, if and only if 𝜏<𝑇.
If the option is exercised at time 𝑄 then the payoff to the borrower is (𝐿𝑄−𝜖𝑉𝑄)+.
If the option is exercised at time 𝑆 then the payoff to the borrower is 𝑎𝑏𝑠(𝐿𝑆−𝜖𝑉𝑆), where abs(.) is the absolute value function.
Notes:
1. If min{𝑄,𝑆}>𝑇 then there is no default option exercise.
2. 𝜖 should be viewed as the recovery rate of the collateral, so ( 1−𝜖) can be viewed as the legal and administrative expenses.
Assume J has intensity 𝜆1 and N has intensity 𝜆2. N is independent of J and W.
Assume the APR of the loan is 𝑅=𝑟0+𝛿𝜆2 where 𝑟0 is the “risk-free” rate, and 𝛿 is a positive parameter to measure the borrower’s creditworthiness in determining the contract rate per period: r.
We have monthly compounding here, so 𝑟=𝑅/12.
Assume that 𝑞𝑡=𝛼+𝛽𝑡, where 𝛽>0, 𝛼<𝑉0/𝐿0 and 𝛽=𝜖−𝛼𝑇.

outputs the default option price, the default probability and the expected exercise time.